+++ HSTracker/HSReplay/HSReplayAPI.swift
@@ -39,6 +39,10 @@ class HSReplayAPI {
                 logger.info("HSReplay: OAuth succeeded")
                 Settings.hsReplayOAuthToken = credential.oauthToken
                 Settings.hsReplayOAuthRefreshToken = credential.oauthRefreshToken
+                AppDelegate.instance().coreManager.exposedHsReplay.setTokens(
+                    accessToken: credential.oauthToken,
+                    refreshToken: credential.oauthRefreshToken
+                )
                 handle()
             },
             failure: { error in
@@ -68,32 +72,6 @@ class HSReplayAPI {
         }
     }
 
-    static func getUploadCollectionToken(handle: @escaping (String) -> Void, failed: @escaping () -> Void) {
-        guard let accountId = MirrorHelper.getAccountId() else {
-            failed()
-            return
-        }
-        oauthswift.startAuthorizedRequest(HSReplay.collectionTokensUrl, method: .GET,
-            parameters: ["account_hi": accountId.hi, "account_lo": accountId.lo], headers: defaultHeaders,
-            onTokenRenewal: tokenRenewalHandler, success: { response in
-            do {
-                guard let json = try response.jsonObject() as? [String: Any], let token = json["url"] as? String else {
-                    logger.error("HSReplay: Unexpected JSON \(String(describing: response.string))")
-                    failed()
-                    return
-                }
-                logger.info("HSReplay : Obtained new collection upload URL")
-                handle(token)
-            } catch {
-                logger.error(error)
-                failed()
-            }
-        }, failure: { error in
-            logger.error(error)
-            failed()
-        })
-    }
-
     static func claimBattleTag(complete: @escaping () -> Void, failed: @escaping () -> Void ) {
         guard let accountId = MirrorHelper.getAccountId(), let battleTag = MirrorHelper.getBattleTag() else {
             failed()
@@ -102,13 +80,7 @@ class HSReplayAPI {
         oauthswift.startAuthorizedRequest("\(HSReplay.claimBattleTagUrl)/\(accountId.hi)/\(accountId.lo)/", method: .POST,
             parameters: ["battletag": battleTag], headers: defaultHeaders,
             onTokenRenewal: tokenRenewalHandler, success: { response in
-            do {
-                let json = try response.jsonObject()
-                logger.info("Claimed battle tag with response \(json)")
-            } catch {
-                logger.error(error)
-                failed()
-            }
+
             complete()
         }, failure: { error in
             logger.error(error)