+++ bootstrap.sh
@@ -1,7 +1,0 @@
-
-
-# Clone translations
-git clone -b macos-test https://github.com/HearthSim/HDT-Localization Translations/
-
-# Download cards
-./scripts/cards_download.sh